+++ CHANGELOG.md
@@ -1,13 +1,24 @@
-## Unreleased
+## 0.0.6
 
 - Add new Profiler middleware
 - Add a copy of the MIT license
+- New debug mode
+- Allow users to pass in their own logger object
 
 **New Metrics**
 
-- `sidekiq.<queue>.<worker>.time`
-- `sidekiq.<queue>.time`
+In `Nightwing::Sidekiq::Profiler` middleware:
+
 - `sidekiq.<queue>.<worker>.retried`
+
+In `Nightwing::Sidekiq::Profiler` middleware:
+
+- `sidekiq.<queue>.time`
+- `sidekiq.<queue>.memory_used`
+- `sidekiq.<queue>.gc.count`
+- `sidekiq.<queue>.<worker>.time`
+- `sidekiq.<queue>.<worker>.gc.count`
+- `sidekiq.<queue>.<worker>.memory_used`
 
 ## 0.0.5